Moving cli code from odb-compiler into its own odb-cli project
