+++ CMakeLists.txt
@@ -80,3 +80,4 @@ add_subdirectory ("thirdparty/googletest")
 add_subdirectory ("odb-sdk")
 add_subdirectory ("odb-compiler")
 add_subdirectory ("odb-interpreter")
+add_subdirectory ("odb-cli")
